Added Data Panel and Data Panel Indicator mounts.
Although there was some small filing necessary, I have added models for
the Data Panel.  In addition a mount for the Data Panel Indicator PCB
was also added.  The DPI mount was siliconed in place, while the DP
mount was just pressure fitted in between the COM8B center support and
bread pan.

Lastly, a servo coupler with larger poles was added.  Not for strength,
but rather to better fill in the holes in the smaller hinges which
ironically were larger than those in the large hinge.